Move cascadeDelete to common.js
