+++ lib/common.js
@@ -17,6 +17,7 @@ require('core-js/es6/string');
 
 var asp = require('rsvp').denodeify;
 var fs = require('graceful-fs');
+var path = require('path');
 var config = require('./config');
 
 // the opposite of extend
@@ -137,3 +138,33 @@ exports.stringify = function (subject) {
   // replace the LF used by `JSON.stringify` with the preferred new line
   return JSON.stringify(subject, null, 2).replace(/\n/g, config.newLine);
 };
+
+/* Recursively remove directory, all those above it, if they are empty.
+ * Takes optional `stopDir` to terminate at. */
+function cascadeDelete(dir, stopDir) {
+  if (dir && dir !== stopDir) {
+    return new Promise(function(resolve) {
+      fs.exists(dir, resolve);
+    })
+    .then(function(exists) {
+      if (!exists)
+        return true;
+
+      return asp(fs.readdir)(dir)
+      .then(function(paths) {
+        if (paths.length === 0) {
+          return asp(fs.rmdir)(dir)
+          .then(function() {
+            return true;
+          });
+        }
+      });
+    })
+    .then(function(cont) {
+      if (cont)
+        return cascadeDelete(path.dirname(dir));
+    });
+  }
+}
+
+exports.cascadeDelete = cascadeDelete;
