+++ lib/install.js
@@ -29,9 +29,10 @@ var globalConfig = require('./global-config');
 
 var rimraf = require('rimraf');
 
-var readJSON = require('./common').readJSON;
 var alphabetize = require('./common').alphabetize;
+var cascadeDelete = require('./common').cascadeDelete;
 var hasProperties = require('./common').hasProperties;
+var readJSON = require('./common').readJSON;
 
 var fs = require('graceful-fs');
 
@@ -868,34 +869,6 @@ function showVersions(forks) {
   }
 }
 exports.showVersions = showVersions;
-
-/* Recursively remove directory, all those above it, if they are empty.
- * Takes optional `stopDir` to terminate at. */
-function cascadeDelete(dir, stopDir) {
-  if (dir && dir !== stopDir) {
-    return new Promise(function(resolve) {
-      fs.exists(dir, resolve);
-    })
-    .then(function(exists) {
-      if (!exists)
-        return true;
-
-      return asp(fs.readdir)(dir)
-      .then(function(paths) {
-        if (paths.length === 0) {
-          return asp(fs.rmdir)(dir)
-          .then(function() {
-            return true;
-          });
-        }
-      });
-    })
-    .then(function(cont) {
-      if (cont)
-        return cascadeDelete(path.dirname(dir));
-    });
-  }
-}
 
 /*
  * Configuration cleaning
